refactor(tests): consolidate messaging tests
